Strip zabbix_valid_server_versions check
This table is incomplete, and a hinderance for deployments to valid
combinations It's also just a guarantee that you'll have code-churn
keeping the table up to date with future operating system releases.

It's better left up to the deployer to know whether their platform
has packages and is supported.

Arguably the zabbix_server_version this should default to 6.0 LTS,
but it's already been set to 6.4, and we don't want to cause further
surprises.