Install MySQL dependencies from mysql task
It's cleaner, less conditional checking, and a single task for all
supported systems.

Remove PyMySQL installation via pip task and switch from mysqldb to
pymysql for all debian based systems. The only reason we're
installing any python/mysql dependencies at all is to use the
ansible community.mysql collection, which has a preference for
pymysql, and mysqldb as a backup [1].

Upgrade pymysql for ubuntu-18.04. Upgrading system-packages with
pip is generally a bad idea, but ubuntu-18.04 comes with
pymysql==0.8.0, which seems to have a problem with newer versions
(>=8) of mysql.

pymysql>=0.9.0,<0.10.0 fixes the issue with passwords missing
during login, but still has a problem setting
log_bin_trust_function_creators.

pymysql>=0.10.0,<0.11.0, fixes the remaining issue. We don't want
to stray to far off from the system default of 0.8.0

[1] https://github.com/ansible-collections/community.mysql/blob/main/plugins/module_utils/mysql.py#L21-L36