Install PostgreSQL dependencies from postresql task
It's cleaner, less conditional checking, and a single task for all
supported systems. The list of packages also remains the same for
all supported systems within the os_family, so we can reduce the
lookup.

The Debian systems had split installing the dependencies in two
tasks, one for the python dependency and one that gets triggered if
zabbix_proxy_install_database_client is true.

We're gonna reuse this variable, and bring it to RedHat systems,
and for mysql aswell in this commit.

And to bring it all back to one task, we use this construction;

- package:
    name: "{{ _dependencies | select | list }}"
  vars:
    _dependencies:
      - "{{ install_client | ternary('client-package', '') }}"
      - some-python-dependency

This will create a list of two items, the python-dependency, and
possibly an empty string. We use `| select` to filter out the empty
strings, and `| list` while strictly not necessary, was
historically appended in case the preceeding result ended up being
a generator.