Refactor PostgreSQL schema import
We take advantage of the community.postgres modules, and do a query
for 'mandatory' in 'dbversion'. If the query fails, we're going to
assume the database has not been populated yet, and rescue the task
by populating it. Thereby alleviating the need for .done files.

While zabbix-server has predictable paths to the schemas, the
zabbix-proxy schema paths are versioned on RedHat based systems.

  e.g. /usr/share/doc/zabbix-proxy-pgsql-X.Y.Z

And you don't necessarily know which version you're installing when
you just want the latest. The code to retain support for older
versions is already in place, by leveraging the ls_output_schema
for the legacy path.