Refactor MySQL schema import
Similar to pgsql, we check for existing dbversion in a block, if
the query fails, we create the database in a rescue section. There
are a few things to note however.

The check and schema creation is not delegated. It happens from
zabbix_proxy. This is to verify that our dbuser can access the
dbname from zabbix_proxy and has all the permissions needed (after
creation) to alter tables between zabbix version upgrades.

For the schema creation we actually need some extra privileges, so
we read what they currently are, set them to what we need, import
our schema, and revert the extra privileges back to their original
state. The reverting happens in the `always:` section, so if any
other task should fail in schema creation (`rescue:` section) we
always revert the privileges before failing.

We also have to deal with varying paths to zabbix-proxy schema on
RedHat based systems.

  e.g. /usr/share/doc/zabbix-proxy-pgsql-X.Y.Z

And you don't necessarily know which version you're installing when
you just want the latest. The code to retain support for older
versions is already in place, by leveraging the ls_output_schema
for the legacy path.