roles/selinux: provide sane defaults
zabbix-server and zabbix-proxy are pretty useless without networking,
and the web interface keeps complaining if it can't connect to zabbix
over port 10050, so might aswell let it.

I really wanted to add httpd_can_network_connect_db, but it's not
needed for the typical "all-in-one" type installation, where you can
stick to unix sockets.