roles/repo: default value for apt priority
Leaving this undefined is just bad form, so might aswell give it
a value. And ff we're actually taking the time to install repo
configuration for the official packages, it's a good indicator that
we also want the packages we install to come from that repository.

P >= 1000 forces these packages even if there are newer available
elsewhere. [1]

[1] https://manpages.debian.org/buster/apt/apt_preferences.5.en.html#How_APT_Interprets_Priorities